test fetching balances

+++ db/migrate/20140118013405_create_transactions.rb
@@ -1,7 +1,7 @@
 class CreateTransactions < ActiveRecord::Migration
   def change
     create_table :transactions do |t|
-      t.integer :amount
+      t.float :amount
       t.belongs_to :account, index: true
 
       t.timestamps
